projects/redhouse-platform/python/libs/rhpy/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rhpy"
version = "0.1.0"
description = "rhpy service components"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"adafruit-blinka>=8.69.0",
	"jsonschema>=4.26.0",
	"logfmter>=0.0.7",
	"netifaces>=0.11.0",
	"prometheus-client>=0.20",
	"pydantic>=2.0",
	"python-dateutil>=2.9.0.post0",
	"redis[hiredis]>=7.3.0",
	"requests>=2.32.5",
]

[project.optional-dependencies]
web = [
	"starlette>=0.37",
	"uvicorn>=0.27",
	"python-multipart>=0.0.9",
]

[project.scripts]
rhpy = "rhpy.main:main"

[tool.setuptools]
package-data = {"rhpy" = ["py.typed"]}

[tool.setuptools.packages.find]
where = ["."]
include = ["rhpy*"]
